+++ spynnaker/pyNN/models/neuron/synapse_dynamics/__init__.py
@@ -29,8 +29,8 @@ from .synapse_dynamics_structural_stdp import SynapseDynamicsStructuralSTDP
 __all__ = ["AbstractGenerateOnMachine", "AbstractPlasticSynapseDynamics",
            "AbstractStaticSynapseDynamics", "AbstractSynapseDynamics",
            "AbstractSynapseDynamicsStructural", "PyNNSynapseDynamics",
+           "SynapseDynamicsStructuralCommon",
            "SynapseDynamicsStatic", "SynapseDynamicsSTDP",
            # Structural plasticity
-           "SynapseDynamicsStructuralCommon",
            "SynapseDynamicsStructuralStatic",
            "SynapseDynamicsStructuralSTDP"]
